Bugfixes in Synchronizer, GameSession and Examples (#14)
I fixed a node count issue in the synchronizer where it wouldn't synch properly without any network nodes.
I also disabled the wait for all nodes to be transferred before doing replication as it seems not to lead to any issues.
I moved the synchronization UI to the basic UI example instead of in the predictive example (was a mistake)
I also put [MDAutoRegister] on the SynchInterface ui class.
Finally I found a bug in the GameSession where on client shutdown it would try to free nodes that were already freed. So I fixed this as well.